A new Matlab wrapper for SwAMP
I added a preliminary function “run_swamp” which can be used to handle
preprocessing tasks before handing off the problem to the mex-ed
binaries. Specifically, it currently handles problem reshaping for the
mean removal case as well as calling the correct binary (swamp, swgamp,
cswgamp) based on the channel type.