An Implementation of Mean Removal for GAMP
It “works,” in that it doesn’t crash, but it does not match the
performance levels of Phil et al (~-18dB NSNR) on arbitrarily large
matrices. I have not yet been able to determine what the issue is. The
framework for the mean removal is equivalent to that used in the AMP
(Gaussian) case.